--- PetSitter/src/main/java/com/cos/blog/TestApplication.java
@@ -1,13 +1,0 @@
-package com.cos.blog;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TestApplication.class, args);
-	}
-
-}
