--- PetSitter/src/main/java/com/cos/blog/dto/ResponseDto.java
@@ -1,15 +1,0 @@
-package com.cos.blog.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ResponseDto<T> {
-	int status;
-	T data;
-}
