--- PetSitter/src/main/java/com/cos/blog/model/RoleType.java
@@ -1,5 +1,0 @@
-package com.cos.blog.model;
-
-public enum RoleType {
-	USER, ADMIN, MANAGER
-}
