+++ PetSitter/src/main/java/com/cos/petsitter/TestApplication.java
@@ -1,0 +1,13 @@
+package com.cos.petsitter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TestApplication.class, args);
+	}
+
+}
