+++ PetSitter/src/main/java/com/cos/petsitter/dto/ResponseDto.java
@@ -1,0 +1,15 @@
+package com.cos.petsitter.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ResponseDto<T> {
+	int status;
+	T data;
+}
