+++ PetSitter/src/main/java/com/cos/petsitter/model/RoleType.java
@@ -1,0 +1,5 @@
+package com.cos.petsitter.model;
+
+public enum RoleType {
+	USER, ADMIN, MANAGER
+}
